LPS-168901 allow node builds for all CSS or JS related client extension
